+++ crates/gpui/src/platform/mac/keyboard.rs
@@ -225,48 +225,3 @@ const TYPEABLE_CODES: &[u16] = &[
     0x002f, // . Period
     0x002c, // / Slash
 ];
-
-#[cfg(test)]
-mod tests {
-    use crate::PlatformKeyboardMapper;
-
-    use super::MacKeyboardMapper;
-
-    #[test]
-    fn test_get_shifted_key() {
-        let mapper = MacKeyboardMapper::new();
-
-        for ch in 'a'..'z' {
-            let key = ch.to_string();
-            let shifted_key = key.to_uppercase();
-            assert_eq!(mapper.get_shifted_key(&key).unwrap(), shifted_key);
-        }
-
-        let shift_pairs = [
-            ("1", "!"),
-            ("2", "@"),
-            ("3", "#"),
-            ("4", "$"),
-            ("5", "%"),
-            ("6", "^"),
-            ("7", "&"),
-            ("8", "*"),
-            ("9", "("),
-            ("0", ")"),
-            ("`", "~"),
-            ("-", "_"),
-            ("=", "+"),
-            ("[", "{"),
-            ("]", "}"),
-            ("\\", "|"),
-            (";", ":"),
-            ("'", "\""),
-            (",", "<"),
-            (".", ">"),
-            ("/", "?"),
-        ];
-        for (key, shifted_key) in shift_pairs {
-            assert_eq!(mapper.get_shifted_key(key).unwrap(), shifted_key);
-        }
-    }
-}
